refactor: 하나의 domain 패키지로 묶기
