+++ src/main/java/baseball/domain/BaseballGame.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 
 import game.Game;
